backend/python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
weaviate-client
chromadb==1.0.20
sentence-transformers>=2.7.0
pinecone
Pillow==10.1.0
transformers==4.36.0
torch>=2.3.0
torchvision>=0.18.0
numpy>=1.24.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.9.2
python-multipart==0.0.6